+++ Assets/Scripts/Helper/TimerManager.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: cd516d0f489aa8f4b8e27d5ca0324e5f
-timeCreated: 1506970006
+guid: 5ad9e5f26ca97724b909a8b228717d32
+timeCreated: 1508472273
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
